add animation force, add run animation control, add normal attack animation control, update physical system, update systems management

+++ Assets/Script/Components/UniqueComponents/GameConfigs/CharacterBaseAttributesComponent.cs
@@ -20,5 +20,5 @@ public class CharacterBaseAttributes
 public class Attributes
 {
     [DataMember]
-    public int baseVelocity;
+    public float baseVelocity;
 }
